requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
# And manually editted ofc

aiohttp==3.11.11
aiomysql>=0.2.0
aiosqlite>=0.21.0
APScheduler>=3.11.0
asyncssh>=2.21.0
bcrypt>=3.2.2
colorama>=0.4.6
httpx>=0.27.0
Hypercorn==0.17.3
lunar_python==1.4.4
mcp>=1.16.0
openai>=2.0
pillow>=12.0
pycryptodome==3.20.0
python-dotenv>=1.0.1
python-magic>=0.4
pytz>=2024.2
Quart==0.20.0
setuptools==70.0.0
tenacity>=9.0.0
websocket==0.2.1
websockets>=12.0
wrapt>=1.17.3
zhconv==1.4.3